refactor: Lotto, LottoSet 구조 변경
