+++ src/main/java/lotto/domain/Lotto.java
@@ -1,8 +1,15 @@
 package lotto.domain;
 
+import camp.nextstep.edu.missionutils.Randoms;
+
 import java.util.List;
+import java.util.stream.Collectors;
 
 public class Lotto {
+    private static int LOTTO_MIN_NUMBER = 1;
+    private static int LOTTO_MAX_NUMBER = 45;
+    private static int LOTTO_ENTITY_SIZE = 6;
+
     private final List<Integer> numbers;
 
     public Lotto(List<Integer> numbers) {
@@ -19,5 +26,14 @@ public class Lotto {
     // TODO: 추가 기능 구현
     public List<Integer> getNumbers(){
         return this.numbers;
+    }
+
+    public static List<Integer> generateLotto(){
+        return Randoms.pickUniqueNumbersInRange(LOTTO_MIN_NUMBER, LOTTO_MAX_NUMBER, LOTTO_ENTITY_SIZE);
+    }
+
+    public static List<Integer> sortLotto(List<Integer> lotto){
+        return lotto.stream().sorted().collect(Collectors.toList());
+
     }
 }
