+++ src/main/java/lotto/domain/LottoSet.java
@@ -6,11 +6,19 @@ import java.util.List;
 public class LottoSet {
 
     private List<Lotto> lottoSet = new ArrayList<>();
-    public LottoSet(List<Lotto> lottoSet){
-        this.lottoSet = lottoSet;
+    public LottoSet(Integer lottoQuantity){
+        this.lottoSet = generateLottoSet(lottoQuantity);
     }
 
     public List<Lotto> getLottoSet() {
         return this.lottoSet;
+    }
+
+    private List<Lotto> generateLottoSet(int lottoQuantity){
+        List<Lotto> lottoSet = new ArrayList<>();
+        for(int i = 0; i < lottoQuantity; i++){
+            lottoSet.add(new Lotto(Lotto.generateLotto()));
+        }
+        return lottoSet;
     }
 }
